requirements.txt
yutto~=2.0.0rc5
httpx~=0.27.2
biliass~=2.1.1
typing_extensions~=4.12.2
pydantic~=2.9.2
colorama~=0.4.6
dict2xml~=1.7.6
aiofiles~=24.1.0
ipython~=8.12.3
traitlets~=5.14.3
uvicorn~=0.32.0
fastapi~=0.115.3
PyYAML~=6.0.2
requests~=2.32.3
schedule~=1.2.2
pandas~=2.2.3
openpyxl~=3.1.5
PyMySQL~=1.1.1
Jinja2~=3.1.4
pyecharts~=2.0.6
simplejson~=3.19.3